Fix `maturin build`
Building `cairo-rs-py` as library requires __not__ linking to the
CPython library. This is achieved by making the default build pass the
`extension-module` feature flag to PyO3 and making it exclusive with
`auto-initialize`.
When testing we do need this to initialize the CPython machine, so it's
a different feature that the `make` target passes to Cargo.